+++ Assets/OxGKit/Utilities/Scripts/Runtime/Timer/DeltaTimer.cs
@@ -1,8 +1,4 @@
-﻿using Cysharp.Threading.Tasks;
-using System;
-using System.Threading;
-
-namespace OxGKit.Utilities.Timer
+﻿namespace OxGKit.Utilities.Timer
 {
     public class DeltaTimer
     {
@@ -34,15 +30,6 @@ namespace OxGKit.Utilities.Timer
 
         ~DeltaTimer()
         {
-        }
-
-        public static async UniTask SetInterval(Action action, int milliseconds, CancellationTokenSource cts = null)
-        {
-            await UniTask.Delay(milliseconds, false, PlayerLoopTiming.Update, (cts == null) ? default : cts.Token);
-
-            action?.Invoke();
-
-            SetInterval(action, milliseconds, cts).Forget();
         }
 
         /// <summary>